+++ classes/Component/AbstractComponent.php
@@ -3,13 +3,10 @@ namespace Component;
 
 use Interfaces\HasBrandInterface;
 use Interfaces\HasNameInterface;
+use Traits\HasNameTrait;
 
 class AbstractComponent implements HasNameInterface, HasBrandInterface
 {
-    /**
-   * @var string
-   */
-    protected $name;
     /**
    * @var string
    */
@@ -22,24 +19,16 @@ class AbstractComponent implements HasNameInterface, HasBrandInterface
         $this->name = $name;
         $this->brand = $brand;
     }
-    //les getters
-    public function getName(): ?string
-    {
-        return  $this->name;
-    }
+    
     public function getBrand(): ?string
     {
         return $this->brand;
     }
-    //les setters
-    public function setName(?string $name): HasNameInterface
-    {
-        $this->name = $name;
-        return $this;
-    }
+    
     public function setBrand($brand): HasBrandInterface
     {
         $this->brand = $brand;
         return $this;
     }
+    use HasNameTrait;
 }
