+++ classes/Devices/AbstractDevice.php
@@ -3,13 +3,10 @@ namespace Devices;
 
 use Interfaces\HasBrandInterface;
 use Interfaces\HasNameInterface;
+use Traits\HasNameTrait;
 
 class AbstractDevice implements HasNameInterface, HasBrandInterface
 {
-    /**
-   * @var string
-   */
-    protected $name;
     /**
    * @var string
    */
@@ -22,21 +19,10 @@ class AbstractDevice implements HasNameInterface, HasBrandInterface
         $this->name = $name;
         $this->brand = $brand;
     }
-    //les getters
-    public function getName(): ?string
-    {
-        return  $this->name;
-    }
+    
     public function getBrand(): ?string
     {
         return $this->brand;
-    }
-    //les setters
-    public function setName(?string $name): HasNameInterface
-    {
-        $this->name = $name;
-
-        return $this;
     }
 
     public function setBrand($brand): HasBrandInterface
@@ -44,4 +30,5 @@ class AbstractDevice implements HasNameInterface, HasBrandInterface
         $this->brand = $brand;
         return $this;
     }
+    use HasNameTrait;
 }
